+++ tests/spikeReport.cpp
@@ -33,6 +33,7 @@
 #include <lunchbox/uint128_t.h>
 
 #define BOOST_TEST_MODULE SpikeReport
+#include <boost/filesystem/operations.hpp>
 #include <boost/filesystem/path.hpp>
 #include <boost/test/unit_test.hpp>